Refactored hierarchy of folders
