+++ training-react/src/app/components/Field/index.js
