+++ training-react/src/app/components/Loading/index.js
