+++ training-react/src/app/components/Loading/layout.js
