+++ training-react/src/app/components/Loading/styles.scss
