+++ training-react/src/app/components/Topbar/styles.scss
