+++ training-react/src/app/screens/Game/components/Board/index.js
