+++ training-react/src/app/screens/Game/components/Board/styles.scss
