+++ training-react/src/app/screens/Game/components/Square/index.js
