+++ training-react/src/app/screens/Game/components/Square/styles.scss
