+++ training-react/src/app/screens/Game/styles.scss
