+++ training-react/src/app/screens/Game/utils.js
