+++ training-react/src/app/screens/User/components/UserInfo/index.js
