+++ training-react/src/app/screens/User/components/UserInfo/styles.scss
