+++ training-react/src/app/screens/User/styles.scss
